# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Муршрут — персональная афиша
  summary: API персональных рекомендаций и готовых маршрутов на вечер
  description: Backend мобильного MAX mini-app «Муршрут». Для защищённых методов сначала получите JWT
    через раздел auth, затем нажмите Authorize.
  version: 1.0.0
paths:
...
  /api/v1/auth/max:
    post:
      tags:
      - auth
      summary: Войти через MAX
      description: Проверяет подпись и срок действия initData и возвращает JWT на 24 часа.
      operationId: auth_api_v1_auth_max_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthInput'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthOut'
        '401':
          description: Требуется действительный Bearer-токен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Запрос не прошёл проверку формата или бизнес-правил.
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/ErrorOut'
                - $ref: '#/components/schemas/RequestValidationErrorOut'
                title: Response 422 Auth Api V1 Auth Max Post
  /api/v1/auth/demo:
    post:
      tags:
      - auth
      summary: Войти в демонстрационный профиль
      description: Доступно только при включённом DEMO_MODE.
      operationId: demo_auth_api_v1_auth_demo_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthOut'
        '404':
          description: Запрошенный ресурс не найден.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
...
  /api/v1/evenings/generate:
    post:
      tags:
      - evenings
      summary: Сгенерировать план вечера
      description: Подбирает новый выполнимый маршрут из двух или трёх событий.
      operationId: generate_api_v1_evenings_generate_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EveningInput'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EveningGenerationOut'
        '401':
          description: Требуется действительный Bearer-токен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '404':
          description: Запрошенный ресурс не найден.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '409':
          description: Операция недоступна до завершения анкеты.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Запрос не прошёл проверку формата или бизнес-правил.
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/ErrorOut'
                - $ref: '#/components/schemas/RequestValidationErrorOut'
                title: Response 422 Generate Api V1 Evenings Generate Post
        '502':
          description: Внешний поставщик событий временно недоступен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
      security:
      - BearerAuth: []
  /api/v1/evenings:
    get:
      tags:
      - evenings
      summary: Получить сохранённые планы
      operationId: saved_api_v1_evenings_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/EveningOut'
                type: array
                title: Response Saved Api V1 Evenings Get
        '401':
          description: Требуется действительный Bearer-токен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
      security:
      - BearerAuth: []
  /api/v1/evenings/{plan_id}:
    get:
      tags:
      - evenings
      summary: Получить план вечера
      operationId: get_api_v1_evenings__plan_id__get
      security:
      - BearerAuth: []
      parameters:
      - name: plan_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Plan Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EveningOut'
        '401':
          description: Требуется действительный Bearer-токен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '404':
          description: Запрошенный ресурс не найден.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Запрос не прошёл проверку формата или бизнес-правил.
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/ErrorOut'
                - $ref: '#/components/schemas/RequestValidationErrorOut'
                title: Response 422 Get Api V1 Evenings  Plan Id  Get
    delete:
      tags:
      - evenings
      summary: Удалить сохранённый план
      operationId: delete_api_v1_evenings__plan_id__delete
      security:
      - BearerAuth: []
      parameters:
      - name: plan_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Plan Id
      responses:
        '204':
          description: Successful Response
        '401':
          description: Требуется действительный Bearer-токен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '404':
          description: Запрошенный ресурс не найден.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Запрос не прошёл проверку формата или бизнес-правил.
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/ErrorOut'
                - $ref: '#/components/schemas/RequestValidationErrorOut'
                title: Response 422 Delete Api V1 Evenings  Plan Id  Delete
  /api/v1/evenings/{plan_id}/save:
    post:
      tags:
      - evenings
      summary: Сохранить план вечера
      operationId: save_api_v1_evenings__plan_id__save_post
      security:
      - BearerAuth: []
      parameters:
      - name: plan_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Plan Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EveningOut'
        '401':
          description: Требуется действительный Bearer-токен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '404':
          description: Запрошенный ресурс не найден.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Запрос не прошёл проверку формата или бизнес-правил.
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/ErrorOut'
                - $ref: '#/components/schemas/RequestValidationErrorOut'
                title: Response 422 Save Api V1 Evenings  Plan Id  Save Post
...
components:
  schemas:
    AuthInput:
      properties:
        init_data:
          type: string
          maxLength: 20000
          title: Init Data
          description: Неизменённая строка initData, полученная от MAX Bridge.
      type: object
      required:
      - init_data
      title: AuthInput
      description: Подписанная строка инициализации MAX WebApp.
      examples:
      - init_data: auth_date=...&user=...&hash=...
    AuthOut:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        user:
          $ref: '#/components/schemas/UserOut'
      type: object
      required:
      - access_token
      - user
      title: AuthOut
      examples:
      - access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        token_type: bearer
        user:
          city: Москва
          first_name: Алексей
          id: 123e4567-e89b-12d3-a456-426614174000
          onboarding_completed: true
...
    ErrorOut:
      properties:
        detail:
          type: string
          title: Detail
      type: object
      required:
      - detail
      title: ErrorOut
      description: Ошибка бизнес-правила или доступа.
    EveningEventOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        category:
          type: string
          title: Category
        category_name:
          type: string
          title: Category Name
        tags:
          items:
            type: string
          type: array
          title: Tags
        image_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Image Url
        image_credit:
          anyOf:
          - type: string
          - type: 'null'
          title: Image Credit
        start_date:
          type: string
          format: date-time
          title: Start Date
        end_date:
          type: string
          format: date-time
          title: End Date
        timezone:
          type: string
          title: Timezone
        price_min:
          anyOf:
          - type: number
          - type: 'null'
          title: Price Min
        price_max:
          anyOf:
          - type: number
          - type: 'null'
          title: Price Max
        is_free:
          type: boolean
          title: Is Free
        city:
          type: string
          title: City
        location_name:
          type: string
          title: Location Name
        address:
          type: string
          title: Address
        latitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Latitude
        longitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Longitude
        source_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Url
        provider:
          type: string
          title: Provider
        reasons:
          items:
            type: string
          type: array
          title: Reasons
        estimated_price:
          anyOf:
          - type: integer
          - type: 'null'
          title: Estimated Price
        next_transfer:
          anyOf:
          - $ref: '#/components/schemas/TransferOut'
          - type: 'null'
      type: object
      required:
      - id
      - title
      - description
      - category
      - category_name
      - tags
      - image_url
      - image_credit
      - start_date
      - end_date
      - timezone
      - price_min
      - price_max
      - is_free
      - city
      - location_name
      - address
      - latitude
      - longitude
      - source_url
      - provider
      - estimated_price
      - next_transfer
      title: EveningEventOut
    EveningGenerationOut:
      properties:
        plan:
          anyOf:
          - $ref: '#/components/schemas/EveningOut'
          - type: 'null'
        reason:
          anyOf:
          - type: string
            enum:
            - no_matches
            - exhausted
          - type: 'null'
          title: Reason
      type: object
      required:
      - plan
      - reason
      title: EveningGenerationOut
    EveningInput:
      properties:
        vibe:
          type: string
          enum:
          - active
          - calm
          - date
          - learn
          - culture
          - surprise
          title: Vibe
        duration_hours:
          type: integer
          enum:
          - 2
          - 4
          - 6
          title: Duration Hours
        budget_max:
          anyOf:
          - type: integer
            enum:
            - 0
            - 1000
            - 3000
          - type: 'null'
          title: Budget Max
        excluded_plan_ids:
          items:
            type: string
            format: uuid
          type: array
          maxItems: 200
          title: Excluded Plan Ids
      type: object
      required:
      - vibe
      - duration_hours
      - budget_max
      title: EveningInput
      description: Условия генерации маршрута на вечер.
      examples:
      - budget_max: 3000
        duration_hours: 4
        excluded_plan_ids: []
        vibe: culture
    EveningOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        saved:
          type: boolean
          title: Saved
        created_at:
          type: string
          format: date-time
          title: Created At
        city:
          type: string
          title: City
        date:
          type: string
          title: Date
        vibe:
          type: string
          title: Vibe
        duration_hours:
          type: integer
          title: Duration Hours
        budget_max:
          anyOf:
          - type: integer
          - type: 'null'
          title: Budget Max
        events:
          items:
            $ref: '#/components/schemas/EveningEventOut'
          type: array
          title: Events
        event_count:
          type: integer
          title: Event Count
        total_cost:
          type: integer
          title: Total Cost
        cost_complete:
          type: boolean
          title: Cost Complete
        duration_minutes:
          type: integer
          title: Duration Minutes
        score:
          type: number
          title: Score
        score_components:
          additionalProperties:
            type: number
          type: object
          title: Score Components
        reasons:
          items:
            type: string
          type: array
          title: Reasons
        warnings:
          items:
            type: string
          type: array
          title: Warnings
      type: object
      required:
      - id
      - saved
      - created_at
      - city
      - date
      - vibe
      - duration_hours
      - budget_max
      - events
      - event_count
      - total_cost
      - cost_complete
      - duration_minutes
      - score
      - score_components
      - reasons
      title: EveningOut
...
    RequestValidationErrorOut:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationIssue'
          type: array
          title: Detail
      type: object
      required:
      - detail
      title: RequestValidationErrorOut
      description: Ошибки автоматической проверки запроса FastAPI.
...
    TransferOut:
      properties:
        distance_km:
          type: number
          title: Distance Km
        minutes:
          type: integer
          title: Minutes
      type: object
      required:
      - distance_km
      - minutes
      title: TransferOut
    UserOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        first_name:
          type: string
          title: First Name
        city:
          type: string
          title: City
        onboarding_completed:
          type: boolean
          title: Onboarding Completed
      type: object
      required:
      - id
      - first_name
      - city
      - onboarding_completed
      title: UserOut
    ValidationIssue:
      properties:
        type:
          type: string
          title: Type
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Loc
        msg:
          type: string
          title: Msg
        input:
          anyOf:
          - {}
          - type: 'null'
          title: Input
        ctx:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Ctx
      additionalProperties: true
      type: object
      required:
      - type
      - loc
      - msg
      title: ValidationIssue
      description: Одна ошибка проверки структуры запроса.
  securitySchemes:
    BearerAuth:
      type: http
      description: JWT, полученный через один из методов авторизации.
      scheme: bearer
    AdminToken:
      type: apiKey
      description: Секрет ручной синхронизации событий.
      in: header
      name: X-Admin-Token
...